More fixes to inventory spacing. This time for long names of eggs and potions.
Was able to contain the text similar to how it is done in the market. Currently, long words (i.e. skeleton) do not center due to the size of images and divs, which still needs to be fixed. More attention will be necessary, but thus should help with Panda Cub Eggs and the like.