Fix for Tavern Chat and extremely long posts
TL;DR
- made tavern chat posts word-wrap, with a max-height to prevent long posts from disrupting scrolling of site

Explanation:

Don't know if anyone noticed that random post by an unnamed person (blank name) with the text of "HELLOOOOO…[repeat a bazillion times].
Well this "broke" chat, meaning there was all of a sudden this side scroll action (which is particularly annoying with multitouch mouse pads). This adjustment prevents such a thing from happening again. Think of "code view" in forums, where text is in it's own little scrolly window.