Hides Rare Pets headline if user has no rare pets
This will hide the rare pets title if the user doesn't have any. 

Debatable change though, maybe we want people to see this and try to find out how they can get a rare pet?

Didn't really test this with a user that had rare pets, is there a quick way to add and remove rare pets from a test user?