Add angular routing choices to header menu
Because tavern and other options can be accessed directly now, these
additions to the menu will make it easier to access features that
currently can take up to 3 clicks to get to.

The code still needs work but at the very least is a start to
implementing issue #1715. I inserted "valueNeeded" in the places I
can't finish.