Re-modify position of markdown down formatting link
I think this one makes more sense than previous PR and it also adjust
spacing of Note. Still interested in feedback (would it make sense to
add to docs.css and call a class or just do it inline?

This is also first time, I've submitted a change to a pull request (and
I'm not sure what's gonna happen). Hopefully github is smart and just
modifies the previous commit (but we'll see).