Fix class guide for new players
The class guide can be activated from the settings page, even when users are too low of a level to set their class. For the most part, this fails gracefully and skips most steps. However, it still shows the equipment tab step, even if the user's level is too low (<4) to see the equipment tab. This temporary tab that the user can't navigate back to through the UI, combined with the tour step saying that their equipment has been modified will alarm many new users.

Fix this by only adding the equipment tour step if drops are enabled for the user (when the inventory tab is enabled).