Cron: construct a query based on options passed to constructCronQuery
So we have different scenarios for running the cron script. This adds
extra options:

- force: Ignore any and all sanity checks on a user's dayStart or
timeZone. Cron WILL be run for all users returned by the mongo
query. Use this when you would like to run cron for missed users or a
missed hour, etc. Defaults to false.
- users: Limit the users returned by the mongo query to those listed in
this array (by _id). Useful for running cron for a specific user,
especially from future admin dashboards for cron. Defaults to undefined.
- currentHour: Finds users where cronTime == currentHour. If not
specified, this will be moment().get('hour') (Server's current hour)
- includeInactives: Whether to include users with >= 0 hp. Usually this
is no, so this defaults to false.
- includeMissed: Whether to include users whose lastCron was over 24
hours ago, even if currentHour is not the same as their
cronTime. Usually this is true, so this defaults to true.