fix(autoAllocate): don't detrain stats on ripening To-Dos
Previously, on cron, the reddening of To-Dos would subtract from the stat training for task-based auto-allocation. This had a tendency to cause users with lots of active tasks to get backward stat allocations on levelup. Now the "down" direction for To-Dos is not counted for attribute training. (Missed Dailies will still detrain.)

Fixes HabitRPG/habitrpg/#2247