chore(VersionError): add debug code to emails showing completed ops
This adds a field to the email if there is an error on batchUpdate. This
field contains all the attempted operations. So the last item in the
list would be the operation that we encountered the error at. @lefnire,
mind taking a look at this and pushing if you feel comfortable with
adding more to the emails?