fix(checklists): Bring MP behavior in sync with XP/Gold
Previously, the player would accrue Mana for every checklist item when a To-Do was completed, regardless of whether or not those checklist items were complete. This was inconsistent with XP and Gold, which were only multiplied based on the number of checklist items previously clicked. Now Mana acts the same way, multiplying MP scored by (1 + completed checklist items).