Pulling in commits from shared repo
Additional changes to other languages are required before merging!

https://trello.com/c/SvTsLdRF/12-translations
"0leg
It would be great if we could have "%number_of% Participants" instead of just "Participants" in the source language so that we could translate it as "участников: %number_of%" which is the only correct way to translate it into Russian without grammatical mistakes. (This example is from Challenges tab.)"
Conflicts:
	locales/en/challenge.json