feat(quests): Allow purchase of quest pets
Implements purchase in the Market of the two currently available pet quest eggs, after the user has completed the corresponding quest at least twice. This is a hard-code solution; @lefnire or others a little more adept in Angular might be able to make this generic and extensible with an ng-repeat--I just couldn't get the expression quite right.