Update npcs.styl
I'm sure it's better to have the popover box visible when the screen gets wider and the columns get near, but still don't collapse to a new row. This small change makes the box visible over the content and not the content visible over the popover box and avoids the problem of going over the header when scrolling.