feat(tooltips): Add expand/collapse cues
Some users (@fragmental) didn't realize that you could click the checklist badge to expand or collapse checklists. This commit adds a tooltip to the badge, as well as to a couple other not-obviously-clickable spots, to cue the user that they can be expanded and collapsed.

BREAKING CHANGE: Will result in "String not found" errors until all locales have an "expandCollapse" string available.