Tour - To-Dos: change "deadline" to "date" (2nd attempt)
Repeat of https://github.com/HabitRPG/habitrpg-shared/pull/207 but without conflicts.

"Deadline" (or even "due date") implies that something happens when the date is reached. "Date" implies that to a lesser degree. We get questions in the Tavern and Newbies Guild about the effect of dates on to-dos, so maybe this is a change worth making.