allow quest progress from attack skills BEFORE quest starts
Currently, if you do Brutal Smash or Burst of Flames BEFORE you start a quest, they do NOT damage the quest boss if you start the quest later in the day (this is the opposite of how tasks work). This also means that if you are not on a boss or collection quest, then those skills can't damage the World Boss. I think this is counterintuitive.

This change causes damage from the skills to be treated just like damage from tasks - it's applied to quests started later that day and to Drag'on.

THIS HAS NOT BEEN TESTED. Sorry. :( I have broken my VM temporarily.