app.js // fix for links (when opening in new page)
Not sure if the above will fix it - still a bit clueless around the Github world. Essentially, the links currently need the '/#' at the beginning in order to work. Using the actual Github software has just caused confusion and crashes, so I am working via the website. If this is not the fix, then somebody who knows better should be able to fix it quick enough.