Change of text if party gems still a thing?
This was a simple change I drafted ages ago, updated to match current text, which may be entirely wrong. Not sure if party gems are still a thing. If not, then this is unnecessary. Again, apologies if none of this makes sense. Still horrendously new and woefully under-qualified. Probably should not even be poking about here. Yay!