Issue #2606 - Verify/fix indices during sortTask
Because the 'sortTask' operation uses a 'from' index and a 'to'
index as parameters, the server needs to check if these indices
are valid before performing the client's request. If the task ID
doesn't match the task at the 'from' index, then it's clear that
the client was looking at a preened list of todos (everything except
more-than-3-days-old-non-quest completed tasks). In that case, try
to recalculate the indices before continuing.