Unlink tasks prior to removing user from challenge.
Per issue #5506 challenge-related tasks are occassionally left on a user after the user is removed from the challenge because the challenge ID (cid) is lost.

To solve this problem attempt to remove challenge-related tasks from user. If this is successful then remove user from list of challenge members. Otherwise fail to error callback.