Move inline styles on task column icons to a class
This uses existing classes instead of adding a new one, because if new
icons are added in the future this should catch them without any extra
classes having to be added to the new icons.

In addition, this will catch any new .option-box added to .task-column
in the future; this may be good or bad, but is easy to fix in any case.