Add alt attributes for images on front page
Some images have an empty alt attribute (alt=""); this isn't a mistake,
it's part of the HTML spec for images that don't add any information.
Debatably, some images marked up as such with this commit do add
information, but that information (e.g. what the Habitica UI looks
like from the intro image) isn't likely to be useful to someone not
able to see.

It's also difficult to convey purely visual concepts like the arrows
with alt-text, especially as the arrows themselves are not images and
therefore cannot have an alt attribute.

Fixes #6560.