Resolves #2433. Adds a text label to exp bar informing user to complete a task to level up when they have enough exp to level.
Media queries for xs screens used to move label below exp bar so it is not cut off.

Refactored jade template to allow the UserCtrl to apply to the .hero-stats div as well, they are all logically contained in the same unit of displaying user info.

The refactor allowed me to replace some of the jade math in the template for displaying HP, EXP, and MP to use the built in functions in statServices service. The benefit is code is being reused rather than having multiple implementations, so refactoring is easier if changes need to be made and lesser chance bugs are introduced. Furthermore it reduces the number of watchers on the header.jade section. Finally, the functions have tests against them and cover additional edge cases for displaying hp, mp and exp.