Store object rather than offset in user store
Previously we were storing the timezone offset, which was causing
problems when trying to know which Timezone the user had selected,
because there are multiple Timezones with the same offset.

We are now storing an object containing the timezone ID and timezone
offset. This allows us to more easily figure out which timezone the
user had selected, and also allows us to pass around both the offset
and the ID without needing to do any additional lookups.

* Move timezoneOffsetToUtc out of computed fields and into methods,
  and allow it to accept an offset
* Add notifcation when saving Timezone
* Add placeholder warning when selected timezone does not match
  system timezone