Store timezone ID for manual timezone
This commit adds a hard-coded ID to each timezone, that will get
stored in the user's preferences. We're storing this ID instead
of the name of the timezone because the name has the potential to
change, and it also limits us in terms of localisation.

Now that we're storing the ID, the offset well need to be looked up
wherever this new manual timezone will be used.

* Add integration test for manual-timezone User endpoint