Fix npm version on Dockerfile to 5.6.0
npm@5.7.1 is having some issue during extraction of large projects (see https://github.com/npm/npm/issues/19989). So while this is not fixed, new contributors were not able to start a new local project, because the containers were updating npm to the newest version. In order to fix this issue I fixed the version to 5.6.0, which is more stable and it works on Habitica project.