remove unused UNSUBSCRIBED code
It looks like it was previously used in cases when a user had left a challenge but there's no need for it any more in normal use, and if it's manually added to a challenge task by an admin, the user is still unable to delete the task.

NOTE that before (if) this is merged, we should check all tasks to ensure it's not used in any of them.