Have separate lowercase version of `heroId`
This both reverts the related unit test back to its original state and creates a constant with a separate, lowercase version of the value passed as the `heroId`. This lowercase version is only used for querying by the `auth.local.lowerCaseUsername` field. The `heroId` itself is still used for finding a valid UUID and for throwing the appropriate error message if no user could be found.

While the previous change to the unit test should have worked, it's still probably best to get back the original username/UUID in the error message instead of the lowercase version.