gear: Fix stats for the "Brilliant Bloom" weapon
This weapon has been introduced in the Summer Splash of 2019.

With an INT bonus of 15 and also being a one-handed weapon, I think it
is way too powerful given that it has a low price and also should be
relatively easy to get.

I've set the stats to match the mage gear for other events and
appropriately made it a two-handed weapon.

Also included small migration script which unequips the shield if a
player happens to have the Brilliant Bloom and a shield.

Signed-off-by: aszlig <aszlig@nix.build>