Fixed disappearing email form bug by forcing a refresh after submission, Fixes #11565  (#11614)
* Fixed disappearing form bug by forcing a refresh after submission

* Added reloadPage() to the changeEmail submit button

* Added indentation for the reloadPage() function

* Adjusted changeUser() function (added reloadPage)

Added reloadPage() function to the changeUser() function. This should allow the page to reload after the email is updated instead of before the function gets to finish updating the account's email

* Adjusted changeUser() function

Added  window.location.reload() to the changeUser() function. This should allow the page to reload after the email is updated instead of before the function gets to finish updating the account's email

* Adjusted changeUser function