Move task icons to the same line as the task name
Especially on smaller screens, space is precious. Having the icons on a
separate line takes up vertical space, when there's plenty of room for
them to sit to the right of the rest of the task content.

bottom: 7px; is to mimic the padding-bottom: 7px; on .task-content.

.task-content is given position: relative in order to make it the
containing block for .icons. Without this, .task-content's parent
.d-flex is the containing block which results in .icons overlapping with
.right-control. .task-content is not affected because no offsets (top,
bottom, left, right) are specified.