Move streak, challenge, and tag icons to the right
Especially on smaller screens, space is precious. Having the icons on a
separate line takes up vertical space, when there's plenty of room for
them to sit to the right of the rest of the task content.

Because the diff is quite messy, here's a list of changes I've made:

 - Make .task-clickable-area a flex-container
 - Add an extra div around .task-clickable-area's current contents to
   keep them from being flexed
 - Move .icons-right from being a child of .icons, to being a child of
   .task-clickable-area
 - Make sure .icons-right stays right-aligned and doesn't shrink, as
   well as keeping its icons bottom-aligned
 - Only show .icons if there is a due date, to shave off those 4px of
   margin when it's not needed
 - Change .task-notes margin-bottom to 4px to account for the fact that
   the margin no longer collapses