Move .icons-right out of .task-clickable-area
I realized that the icons were no longer vertically aligned with the due
date, and this became even more obvious when a checklist was added. In
order to keep them aligned, I moved them out of .task-clickable-area and
matched its padding-bottom.

List of changes (messy diff):

 - Remove .d-flex from .task-clickable-area
 - Remove extra div around .task-clickable-area's contents
 - Move .icons-right from being a child of .task-clickable-area, to
   being a child of its parent element

Remaining issues are the border on :hover around the task box, and the options button.