Move <menu-dropdown> to the right in .icons-right
After .icons-right was moved outside of .task-clickable-area, the
options dropdown button was no longer in the top right of the task box
because .icons-right was to the right of it.

List of changes (messy diff):

 - Move <menu-dropdown> from being a child of the parent of .task-title
   to being a child of .icons-right
 - Add a <div> around all of the icons in .icon-right to keep them in a
   row
 - Add padding-top to .icons-right to match .task-clickable-area so the
   dropdown button is in (almost) the same position
 - Remove the margin-top from .icons-right, for the above purpose

<menu-dropdown>'s new position is 1px higher than its previous one,
probably because of some strange layout rules from being inside an extra
div with .task-title.