update profile and profileModal from comments
I made the changes with the comments but this still doesn't seem to work. If this is not just a issue on my end I still think its some sort of rendering issue with the browser and router. I checked to see if document.title does get updated after the modal gets hidden and it does update the title but browser doesn't change.