make changes to api v3 test case
I changed the test for api v3 because I changed what mark-pms-read does and its not producing the expected result. I will try and see if I can make the test work so that the test passes. for now I changed it so that expected is set to 1 instead of 0.