revert Hall of Heroes to version from develop
The changes made to this file for the Admin Panel were optional
and only for slight convenience - they allowed an admin to
go to the Hall of Heroes and open any user listed there
into the Admin Panel by clicking a link.
The same result can be achieved by manually copying the UUID or Username
and pasting into the Admin Panel.

Now that the Hall of Heroes has been modified in `develop`,
we can just use that version. There's no real point to
updating it for the optional Admin Panel change,
especially since the bulk of the Hall's admin code will be
removed after the Admin Panel goes live.