Update gear.json
[//]: # (Note: See https://habitica.fandom.com/wiki/Using_Your_Local_Install_to_Modify_Habitica%27s_Website_and_API for more info)

[//]: # (Put Issue # here, if applicable. This will automatically close the issue if your PR is merged in)
Fixes put_#_and_issue_number_here

### Changes
[//]: # (Describe the changes that were made in detail here. Include pictures if necessary)
Added shieldSpecialSummer2022RogueText and shieldSpecialSummer2022RogueNotes, based on in-game name & description.

Important note! I took the description for the "Notes" field from the in-game description, but it's identical to the one for the Rogue weapon, down to and including it being a Strength item, not a Constitution item.  I'm not sure whether that's accurate, or if it's just pulling from the weapon description since the shield description isn't available.

(Sorry if I shouldn't have done this without being certain of the correct description -- last time I reported a bug in gear.json I was encouraged to just do a pull request next time, so this is my first-ever pull request.)

[//]: # (Put User ID in here - found on the Habitica website at User Icon > Settings > API)

----
UUID: 6fa2ed29-9913-49c7-a089-2c16f1713a05